Improve error handling, API consistency, and tests
- Remove Console.WriteLine; use exceptions for fallback logic
- OrderEvaluator: throw InvalidOperationException for duplicate OrderBy/OrderByDescending, with clear error messages
- IEvaluator: return non-nullable IQueryable<T> for consistency
- Builder Count properties now use int, not long
- Query.Clear() resets all properties, including pagination
- SetPage and Take property now validate input and throw with details
- Add and improve XML documentation for public APIs
- Query<T>.ToString() provides detailed debug info
- Add comprehensive unit tests for edge cases and validation
- Codebase is now more robust, maintainable, and production-ready

+++ 1. Source/2. Data Access/Craft.QuerySpec/Builders/EntityFilterBuilder.cs
@@ -6,6 +6,9 @@ namespace Craft.QuerySpec;
 /// <summary>
 /// Builder for creating and managing a list of entity filter criteria for type <typeparamref name="T"/>.
 /// </summary>
+/// <remarks>
+/// This class is NOT thread-safe. Do not share instances across threads.
+/// </remarks>
 /// <typeparam name="T">The entity type.</typeparam>
 [Serializable]
 public class EntityFilterBuilder<T> where T : class
@@ -23,7 +26,7 @@ public class EntityFilterBuilder<T> where T : class
     /// <summary>
     /// Gets the number of filter criteria.
     /// </summary>
-    public long Count => EntityFilterList.Count;
+    public int Count => EntityFilterList.Count;
 
     /// <summary>
     /// Adds a filter criteria based on a boolean expression.
